Update type & add copyright

+++ models/error.go
@@ -1,4 +1,5 @@
 // Copyright 2015 The Gogs Authors. All rights reserved.
+// Copyright 2019 The Gitea Authors. All rights reserved.
 // Use of this source code is governed by a MIT-style
 // license that can be found in the LICENSE file.
 
@@ -881,7 +882,7 @@ func (err ErrIssueNotExist) Error() string {
 type ErrIssueInvalidPriority struct {
 	ID              int64
 	RepoID          int64
-	DesiredPriority int
+	DesiredPriority int64
 }
 
 // IsErrIssueInvalidPriority checks if an error is a ErrIssueInvalidPriority.
